+++ src/main/java/dev/willyelton/crystal_tools/common/levelable/block/entity/CrystalQuarryBlockEntity.java
@@ -113,6 +113,12 @@ public class CrystalQuarryBlockEntity extends LevelableBlockEntity implements Me
     }
 
     // TODO
+    @Override
+    protected void resetExtraSkills() {
+
+    }
+
+    // TODO
     protected final ContainerData dataAccess = new LevelableContainerData(this) {
         @Override
         protected int getExtra(int index) {