+++ src/main/java/dev/willyelton/crystal_tools/common/network/RegisterPackets.java
@@ -10,6 +10,7 @@ import dev.willyelton.crystal_tools.common.network.data.ModeSwitchPayload;
 import dev.willyelton.crystal_tools.common.network.data.OpenBackpackPayload;
 import dev.willyelton.crystal_tools.common.network.data.RemoveItemPayload;
 import dev.willyelton.crystal_tools.common.network.data.RemoveXpPayload;
+import dev.willyelton.crystal_tools.common.network.data.ResetSkillsBlockPayload;
 import dev.willyelton.crystal_tools.common.network.data.ResetSkillsPayload;
 import dev.willyelton.crystal_tools.common.network.data.ScrollPayload;
 import dev.willyelton.crystal_tools.common.network.data.SkillCacheUpdatePayload;
@@ -25,6 +26,7 @@ import dev.willyelton.crystal_tools.common.network.handler.ModeSwitchHandler;
 import dev.willyelton.crystal_tools.common.network.handler.OpenBackpackHandler;
 import dev.willyelton.crystal_tools.common.network.handler.RemoveItemHandler;
 import dev.willyelton.crystal_tools.common.network.handler.RemoveXpHandler;
+import dev.willyelton.crystal_tools.common.network.handler.ResetSkillsBlockHandler;
 import dev.willyelton.crystal_tools.common.network.handler.ResetSkillsHandler;
 import dev.willyelton.crystal_tools.common.network.handler.ScrollHandler;
 import dev.willyelton.crystal_tools.common.network.handler.SkillCacheUpdateHandler;
@@ -57,6 +59,7 @@ public class RegisterPackets {
         registrar.playToServer(OpenBackpackPayload.TYPE, OpenBackpackPayload.STREAM_CODEC, OpenBackpackHandler.INSTANCE::handle);
         registrar.playToServer(RemoveXpPayload.TYPE, RemoveXpPayload.STREAM_CODEC, RemoveXpHandler.INSTANCE::handle);
         registrar.playToServer(VeinMiningPayload.TYPE, VeinMiningPayload.STREAM_CODEC, VeinMiningHandler.INSTANCE::handle);
+        registrar.playToServer(ResetSkillsBlockPayload.TYPE, ResetSkillsBlockPayload.STREAM_CODEC, ResetSkillsBlockHandler.INSTANCE::handle);
 
         // Server to Client
         registrar.playToClient(SkillCacheUpdatePayload.TYPE, SkillCacheUpdatePayload.STREAM_CODEC, SkillCacheUpdateHandler.INSTANCE::handle);